+++ Extensions/Minded.Extensions.Configuration/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Reflection;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Minded.Extensions.Configuration
@@ -15,24 +16,10 @@ namespace Minded.Extensions.Configuration
         /// <param name="serviceCollection">Service Collection used to control the dependency injection</param>
         /// <param name="assemblyFilter">Filter function used to select the assemblies to scan</param>
         /// <param name="MindedBuilder"><paramref name="MindedBuilder"/></param>
-        public static void AddMinded(this IServiceCollection serviceCollection, Func<AssemblyName, bool> assemblyFilter = null,
+        public static void AddMinded(this IServiceCollection serviceCollection, IConfiguration configuration, Func<AssemblyName, bool> assemblyFilter = null,
             Action<MindedBuilder> MindedBuilder = null)
         {
-            var builder = new MindedBuilder(serviceCollection, assemblyFilter);
-
-            // Register Mediators
-            //builder.RegisterMediator();
-
-            //// Register the Validators
-            //builder.RegisterValidators();
-
-            //// Register all the query handlers with the related decorators
-            //builder.RegisterQueryHandlers();
-
-            //// Register all the command handlers with the related decorators
-            //builder.RegisterCommandHandlers();
-
-            
+            var builder = new MindedBuilder(serviceCollection, configuration, assemblyFilter);
             MindedBuilder?.Invoke(builder);
         }
     }