+++ Extensions/Minded.Extensions.Logging/Decorator/ServiceCollectionExtensions.cs
@@ -10,23 +10,23 @@ namespace Minded.Extensions.Logging.Decorator
         /// <summary>
         /// Add the Logging decorator for Commands
         /// </summary>
-        public static MindedBuilder AddCommandLoggingDecorator(this MindedBuilder builder, IConfiguration configuration)
+        public static MindedBuilder AddCommandLoggingDecorator(this MindedBuilder builder)
         {
             builder.QueueCommandDecoratorRegistrationAction((b, i) => b.DecorateHandlerDescriptors(i, typeof(LoggingCommandHandlerDecorator<>)));
             builder.QueueCommandWithResultDecoratorRegistrationAction((b, i) => b.DecorateHandlerDescriptors(i, typeof(LoggingCommandHandlerDecorator<,>)));
 
-            builder.ServiceCollection.Configure<LoggingOptions>(configuration.GetSection("Minded:LoggingOptions"));
+            builder.ServiceCollection.Configure<LoggingOptions>(builder.Configuration.GetSection("Minded:LoggingOptions"));
             return builder;
         }
 
         /// <summary>
         /// Add the Logging decorator for Queries
         /// </summary>
-        public static MindedBuilder AddQueryLoggingDecorator(this MindedBuilder builder, IConfiguration configuration)
+        public static MindedBuilder AddQueryLoggingDecorator(this MindedBuilder builder)
         {
             builder.QueueQueryDecoratorRegistrationAction((b, i) => b.DecorateHandlerDescriptors(i, typeof(LoggingQueryHandlerDecorator<,>)));
 
-            builder.ServiceCollection.Configure<LoggingOptions>(configuration.GetSection("Minded:LoggingOptions"));
+            builder.ServiceCollection.Configure<LoggingOptions>(builder.Configuration.GetSection("Minded:LoggingOptions"));
             return builder;
         }
     }