+++ Extensions/Minded.Extensions.WebApi/RestMediator.cs
@@ -16,19 +16,19 @@ namespace Minded.Extensions.WebApi
             _rulesProcessor = rulesProcessor;
         }
 
-        public async Task<ActionResult> ProcessRestQueryAsync<TResult>(RestOperation operation, IQuery<TResult> query)
+        public async Task<IActionResult> ProcessRestQueryAsync<TResult>(RestOperation operation, IQuery<TResult> query)
         {            
             var result = await ProcessQueryAsync(query);
             return _rulesProcessor.ProcessQueryRules(operation, result);
         }
 
-        public async Task<ActionResult> ProcessRestCommandAsync(RestOperation operation, ICommand command)
+        public async Task<IActionResult> ProcessRestCommandAsync(RestOperation operation, ICommand command)
         {
             var result = await ProcessCommandAsync(command);
             return _rulesProcessor.ProcessCommandRules(operation, result);
         }
 
-        public async Task<ActionResult> ProcessRestCommandAsync<TResult>(RestOperation operation, ICommand<TResult> command)
+        public async Task<IActionResult> ProcessRestCommandAsync<TResult>(RestOperation operation, ICommand<TResult> command)
         {
             var result = await ProcessCommandAsync<TResult>(command);
             return _rulesProcessor.ProcessCommandRules(operation, result);